Fix Spamming User Icon Leaving Group
There's some issue with the CSS classes not getting properly imported from GroupForum.css. When I change the values in there, the classes don't update. However, if I switched them to the classes used by the courses, it did work, and switching all of them made the problem go away.